[#72661] Namespace Backlogs Rails routes
Move the Rails-facing Backlogs page and sprint endpoints under the
Backlogs namespace and standardize their route helpers.

This renames the legacy `rb_*` page controllers, moves the canonical
backlog details route under `/backlogs/backlog`, and updates the touched
sprint endpoints to use more conventional controller/action naming.

It leaves the deeper `OpenProject::Backlogs` library internals (patches,
API endpoints, hooks) untouched so this stays a Rails-edge rename.

https://community.openproject.org/wp/72661
