+++ app/seeders/demo_data/references.rb
@@ -48,8 +48,8 @@ module DemoData
     #
     # For instance:
     # - Turns `##sprint:sprint_backlog` into
-    #   `/projects/demo-project/sprints/23/taskboard` given there is a sprint
-    #   referenced with :sprint_backlog and its ID here is 23.
+    #   `/projects/demo-project/backlogs/sprints/23/taskboard` given there is a
+    #   sprint referenced with :sprint_backlog and its ID here is 23.
     #
     #   Alternatively `##sprint.id:sprint_backlog` is translated into just the
     #   id.
@@ -123,7 +123,7 @@ module DemoData
     end
 
     def sprint_link(sprint)
-      url_helpers.backlogs_project_sprint_taskboard_path(
+      url_helpers.project_backlogs_sprint_taskboard_path(
         sprint_id: sprint.id,
         project_id: sprint.project.identifier
       )
