+++ modules/backlogs/app/components/backlogs/inbox_component.html.erb
@@ -99,7 +99,7 @@ See COPYRIGHT and LICENSE files for more details.
               block: true,
               align_content: :start,
               underline: true,
-              href: backlog_backlogs_project_backlogs_path(project, all: 1)
+              href: project_backlogs_backlog_path(project, all: 1)
             )
           ) { t(".show_more", count: middle_count) }
         %>
