+++ modules/backlogs/app/components/backlogs/inbox_item_component.html.erb
@@ -52,7 +52,7 @@ See COPYRIGHT and LICENSE files for more details.
       <%= render(
             Primer::Alpha::ActionMenu.new(
               menu_id: dom_target(work_package, :menu),
-              src: menu_project_inbox_path(project, work_package),
+              src: menu_project_backlogs_inbox_path(project, work_package),
               anchor_align: :end,
               classes: "hide-when-print"
             )
