+++ modules/backlogs/app/components/backlogs/inbox_item_component.rb
@@ -64,11 +64,11 @@ module Backlogs
         data: {
           draggable_id: work_package.id,
           draggable_type: "story",
-          drop_url: move_project_inbox_path(project, work_package),
+          drop_url: move_project_backlogs_inbox_path(project, work_package),
           story: true,
           controller: "backlogs--story",
           backlogs__story_id_value: work_package.id,
-          backlogs__story_split_url_value: details_backlogs_project_backlogs_path(project, work_package),
+          backlogs__story_split_url_value: project_backlogs_backlog_details_path(project, work_package),
           backlogs__story_full_url_value: work_package_path(work_package),
           backlogs__story_selected_class: "Box-row--blue",
           test_selector: card_test_selector
