+++ modules/backlogs/app/components/backlogs/move_to_sprint_dialog_component.html.erb
@@ -33,7 +33,7 @@ See COPYRIGHT and LICENSE files for more details.
 
     d.with_body do
       primer_form_with(
-        url: move_project_inbox_path(project, work_package),
+        url: move_project_backlogs_inbox_path(project, work_package),
         method: :put,
         id: FORM_ID,
         data: { turbo_stream: true }
