+++ modules/backlogs/app/components/backlogs/story_menu_list_component.html.erb
@@ -33,7 +33,7 @@ See COPYRIGHT and LICENSE files for more details.
       id: dom_target(story, :menu, :open_details),
       tag: :a,
       label: t(:"js.button_open_details"),
-      href: details_backlogs_project_backlogs_path(project, story),
+      href: project_backlogs_backlog_details_path(project, story),
       content_arguments: { data: { turbo_frame: "content-bodyRight", turbo_action: "advance" } }
     ) do |item|
       item.with_leading_visual_icon(icon: :"op-view-split")
