+++ modules/backlogs/app/controllers/backlogs/taskboard_controller.rb
@@ -1,3 +1,5 @@
+# frozen_string_literal: true
+
 #-- copyright
 # OpenProject is an open source project management software.
 # Copyright (C) the OpenProject GmbH
@@ -26,32 +28,16 @@
 # See COPYRIGHT and LICENSE files for more details.
 #++
 
-require "spec_helper"
+module Backlogs
+  class TaskboardController < ::RbApplicationController
+    menu_item :backlogs
 
-RSpec.describe RbMasterBacklogsController do
-  describe "routing" do
-    it {
-      expect(get("/projects/project_42/backlogs")).to route_to(controller: "rb_master_backlogs",
-                                                               action: "index",
-                                                               project_id: "project_42")
-    }
+    def show
+      @board = @sprint.task_board_for(@project)
 
-    it {
-      route = "/projects/project_42/backlogs/backlog"
-      expect(get(route)).to route_to(controller: "rb_master_backlogs",
-                                     action: "backlog",
-                                     project_id: "project_42")
-    }
+      return redirect_to(project_work_package_board_path(@project, @board)) if @board
 
-    it {
-      expect(get("/projects/project_42/backlogs/details/33")).to route_to(
-        controller: "rb_master_backlogs",
-        action: "details",
-        project_id: "project_42",
-        work_package_id: "33",
-        tab: :overview,
-        work_package_split_view: true
-      )
-    }
+      render_404
+    end
   end
 end
