+++ modules/backlogs/app/views/backlogs/backlog/show.html.erb
@@ -30,7 +30,7 @@ See COPYRIGHT and LICENSE files for more details.
 <% html_title t(:label_backlogs) %>
 
 <% content_controller "backlogs",
-                      "backlogs-list-url-value": backlogs_project_backlogs_path(@project) %>
+                      "backlogs-list-url-value": project_backlogs_backlog_path(@project) %>
 
 <% content_for :content_header do %>
   <%=
@@ -38,7 +38,7 @@ See COPYRIGHT and LICENSE files for more details.
       header.with_title { t(:label_backlog_and_sprints) }
       header.with_breadcrumbs(
         [{ href: project_overview_path(@project), text: @project.name },
-         { href: backlog_backlogs_project_backlogs_path(@project), text: t(:label_backlogs) },
+         { href: project_backlogs_backlog_path(@project), text: t(:label_backlogs) },
          t(:label_backlog_and_sprints)]
       )
     end
@@ -48,7 +48,7 @@ See COPYRIGHT and LICENSE files for more details.
 <% content_for :content_body do %>
   <%= turbo_frame_tag :backlogs_container,
                       refresh: :morph,
-                      src: backlog_backlogs_project_backlogs_path(@project, all: show_all_backlog),
+                      src: project_backlogs_backlog_path(@project, all: show_all_backlog),
                       class: "op-backlogs-page" %>
 <% end %>
