+++ modules/backlogs/config/locales/en.yml
@@ -201,6 +201,11 @@ en:
         copy_work_package_id: "Copy work package ID"
         move_menu: "Move"
 
+    burndown_charts:
+      show:
+        blankslate_title: "No burndown data available"
+        blankslate_description: "Set start and end date for the sprint to generate a burndown chart."
+
   burndown:
     story_points: "Story points"
     story_points_ideal: "Story points (ideal)"
@@ -253,9 +258,5 @@ en:
             caption: "Sprints created in this project will be available to all subprojects of the current project."
             info: "Sharing a sprint will share the name, status and the start and finish dates in all projects. These cannot be modified in projects that receive and use these sprints."
         sprint_sharing: Share sprints
-  rb_burndown_charts:
-    show:
-      blankslate_title: "No burndown data available"
-      blankslate_description: "Set start and end date for the sprint to generate a burndown chart."
 
   remaining_hours: "remaining work"
