+++ modules/backlogs/spec/features/backlogs/create_spec.rb
@@ -52,7 +52,7 @@ RSpec.describe "Create", :js do
 
     within ".PageHeader-breadcrumbs" do
       expect(page).to have_link(href: project_path(project), text: project.name)
-      expect(page).to have_link(href: backlog_backlogs_project_backlogs_path(project), text: "Backlogs")
+      expect(page).to have_link(href: project_backlogs_backlog_path(project), text: "Backlogs")
       expect(page).to have_text("Backlog and sprints")
     end
   end
