+++ modules/backlogs/spec/routing/backlogs/stories_routing_spec.rb
@@ -30,11 +30,11 @@
 
 require "spec_helper"
 
-RSpec.describe RbStoriesController do
+RSpec.describe Backlogs::StoriesController do
   describe "routing" do
     it {
-      expect(put("/projects/project_42/sprints/21/stories/85/move")).to route_to(
-        controller: "rb_stories",
+      expect(put("/projects/project_42/backlogs/sprints/21/stories/85/move")).to route_to(
+        controller: "backlogs/stories",
         action: "move",
         project_id: "project_42",
         sprint_id: "21",
@@ -43,8 +43,8 @@ RSpec.describe RbStoriesController do
     }
 
     it {
-      expect(post("/projects/project_42/sprints/21/stories/85/reorder")).to route_to(
-        controller: "rb_stories",
+      expect(post("/projects/project_42/backlogs/sprints/21/stories/85/reorder")).to route_to(
+        controller: "backlogs/stories",
         action: "reorder",
         project_id: "project_42",
         sprint_id: "21",
@@ -53,8 +53,8 @@ RSpec.describe RbStoriesController do
     }
 
     it {
-      expect(get("/projects/project_42/sprints/21/stories/85/menu")).to route_to(
-        controller: "rb_stories",
+      expect(get("/projects/project_42/backlogs/sprints/21/stories/85/menu")).to route_to(
+        controller: "backlogs/stories",
         action: "menu",
         project_id: "project_42",
         sprint_id: "21",
