+++ modules/backlogs/spec/views/backlogs/burndown_charts/show_spec.rb
@@ -30,7 +30,7 @@
 
 require "spec_helper"
 
-RSpec.describe "rb_burndown_charts/show" do
+RSpec.describe "backlogs/burndown_charts/show" do
   let(:user1) { create(:user) }
   let(:user2) { create(:user) }
   let(:role_allowed) do
@@ -115,7 +115,7 @@ RSpec.describe "rb_burndown_charts/show" do
       render
 
       expect(view).to render_template(partial: "_burndown", count: 0)
-      expect(rendered).to include(I18n.t("rb_burndown_charts.show.blankslate_title"))
+      expect(rendered).to include(I18n.t("backlogs.burndown_charts.show.blankslate_title"))
     end
   end
 end
